Add operator do no kNN search and get best matchin hemispheres
Also add functions to Hemisphere object to simplify computation
of variables of interest. Add selector perform the mixing, for the
moment the output is still not written out. The knn closest hemispheres
are saved to Event in memory. An additional module will take
charge of writting out mixed events (possibly several combinations).