Use TEventList to select subsets of TChain for creating pseudodata
Created a Python helper class to produce TEventList that can be
associated with TChains and the used to only go through a subset
of the events. So far the event list takes the first N events
per subsample but it can be randomized. Also now writting the
original pseudodata in addition to the mixed one (in the same
file).